+++ render_patterns.api.php
@@ -1,10 +1,8 @@
 <?php
+
 /**
  * @file
  * Defines the API functions provided by the render_patterns module.
- *
- * @ingroup render_patterns
- * @{
  */
 
 /**
@@ -12,33 +10,32 @@
  *
  * Provide information of how your module or theme is using this module.
  *
- * @return  array
+ * @return array
  *   - directory String The directory wherein to search for classes.  They will
  *   be autoloaded.
  *   - weight int Optional.  When omitted your module's system weight is used.
  *     This value is used in the autoloader registry.
  *
- * @see hook_render_patterns_info_alter().
+ * @see hook_render_patterns_info_alter()
  */
 function hook_render_patterns_info() {
-  return array(
+  return [
     'directory' => drupal_get_path('module', 'my_module') . '/render_patterns',
     // Give the theme the higher weight so it can override.
     'weight' => 10,
-  );
+  ];
 }
 
 /**
  * Implements hook_render_patterns_info_alter().
  *
- * @see  hook_render_patterns_info().
+ * @see hook_render_patterns_info()
  */
 function hook_render_patterns_info_alter(&$info) {
-  // Chenge the default location of the render_patterns dir for the default theme.
-  // @FIXME
-// // @FIXME
-// // This looks like another module's variable. You'll need to rewrite this call
-// // to ensure that it uses the correct configuration object.
-// $info['render_patterns']['directory'] = drupal_get_path('theme', variable_get('theme_default', '')) . '/includes/patterns';
 
+  // Change the location of the render_patterns for the default theme.
+  $default_theme = \Drupal::service('config.factory')
+    ->get('system.theme')
+    ->get('default');
+  $info['render_patterns']['directory'] = \Drupal::root() . '/' . drupal_get_path('theme', $default_theme) . '/includes/patterns';
 }