+++ src/Uncacheable.php
@@ -3,9 +3,12 @@
 namespace Drupal\render_patterns;
 
 /**
- * A class to hold a cacheable value.
+ * A class to hold an uncacheable value.
+ *
+ * Wrap default values returned by default__* methods in this and the data will
+ * not be cached.
  */
-class Cacheable {
+class Uncacheable {
 
   /**
    * The cacheable value.
@@ -20,10 +23,10 @@ class Cacheable {
    * @param mixed $value
    *   An arbitrary value that may be cached by the pattern.
    *
-   * @return \Drupal\render_patterns\Cacheable
+   * @return \Drupal\render_patterns\Uncacheable
    *   A new instance with embedded $value.
    */
-  public static function value($value): Cacheable {
+  public static function value($value): Uncacheable {
     $instance = new static();
     $instance->value = $value;
     return $instance;